requirements.txt
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
coverage==7.2.1
cryptography==39.0.1
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.92.0
flake8==5.0.4
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.5
pyflakes==2.5.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-cov==3.0.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
websockets==10.4
